WIP: prepare to swap pure java impl for Hdf5DataProcessingReader
