+++ vcell-core/src/main/java/cbit/vcell/simdata/SimulationData.java
@@ -981,7 +981,7 @@ public synchronized SimDataBlock getSimDataBlock(OutputContext outputContext, St
 			PDEDataInfo pdeDataInfo = new PDEDataInfo(vcDataId.getOwner(),vcDataId.getID(),varName,time,lastDataProcessingOutputInfoTime);
 			DataProcessingOutputDataValuesOP dataProcessingOutputDataValuesOP =
 				new DataProcessingOutputDataValuesOP(vcDataId, varName, TimePointHelper.createSingleTimeTimePointHelper(extractClosestPostProcessTime(time)), DataIndexHelper.createAllDataIndexesDataIndexHelper(), outputContext, null);
-			Hdf5DataProcessingReader hdf5DataProcessingReader = new Hdf5DataProcessingReader();
+			Hdf5DataProcessingReaderNative hdf5DataProcessingReader = new Hdf5DataProcessingReaderNative();
 			DataProcessingOutputDataValues dataProcessingOutputDataValues =
 				(DataProcessingOutputDataValues) hdf5DataProcessingReader.getDataProcessingOutput(dataProcessingOutputDataValuesOP, getDataProcessingOutputSourceFileHDF5());
 			return new SimDataBlock(pdeDataInfo,dataProcessingOutputDataValues.getDataValues()[0]/*1 time only*/,VariableType.POSTPROCESSING);
@@ -1083,7 +1083,7 @@ synchronized double[][][] getSimDataTimeSeries0(
 				DataProcessingOutputDataValuesOP dataProcessingOutputDataValuesOP =
 					new DataProcessingOutputDataValuesOP(vcDataId, varNames[i],
 						TimePointHelper.createSpecificTimePointHelper(specificTimePoints), DataIndexHelper.createSpecificDataIndexHelper(indexes[i]), outputContext, null);
-				Hdf5DataProcessingReader hdf5DataProcessingReader = new Hdf5DataProcessingReader();
+				Hdf5DataProcessingReaderNative hdf5DataProcessingReader = new Hdf5DataProcessingReaderNative();
 				DataProcessingOutputDataValues dataProcessingOutputDataValues =
 					(DataProcessingOutputDataValues) hdf5DataProcessingReader.getDataProcessingOutput(dataProcessingOutputDataValuesOP, getDataProcessingOutputSourceFileHDF5());
 				for (int j = 0; j < specificTimePoints.length; j++) {
@@ -1811,7 +1811,7 @@ private void refreshDataProcessingOutputInfo(OutputContext outputContext) throws
 		lastOutputContext = outputContext;
 		DataProcessingOutputInfoOP dataProcessingOutputInfoOP = new DataProcessingOutputInfoOP(vcDataId, false, outputContext);
 		try{
-			Hdf5DataProcessingReader hdf5DataProcessingReader = new Hdf5DataProcessingReader();
+			Hdf5DataProcessingReaderNative hdf5DataProcessingReader = new Hdf5DataProcessingReaderNative();
 			dataProcessingOutputInfo = (DataProcessingOutputInfo) hdf5DataProcessingReader.getDataProcessingOutput(dataProcessingOutputInfoOP, dataProcessingOutputFile);
 		}catch(Exception e){
 			throw new DataAccessException(e.getMessage(),(e.getCause()==null?e:e.getCause()));
